KNOX-2474 - RemoteConfigurationRegistryJAASConfigTest fails due to invalid auth
